+++ extensions/adb/src/adb-open-url.ts
@@ -1,12 +1,13 @@
 import { LaunchProps, showHUD } from "@raycast/api";
 import { execSync } from "child_process";
+import checkAdbExists from "./utils";
 
 interface AdbUrlArguments {
   url: string;
 }
 
 export default async function openUrl(props: LaunchProps<{ arguments: AdbUrlArguments }>) {
-  const adbDir = `${process.env.HOME}/Library/Android/sdk/platform-tools/adb`;
+  const adbDir = await checkAdbExists();
   const url = props.arguments.url;
   await showHUD("️🌐 Opening " + url);
   execSync(`${adbDir} shell am start -a android.intent.action.VIEW -d  '${url}'`);