+++ extensions/adb/src/adb-screenshots.ts
@@ -2,10 +2,11 @@ import { Clipboard, showHUD } from "@raycast/api";
 import moment from "moment";
 import { execSync } from "child_process";
 import * as fs from "fs";
+import checkAdbExists from "./utils";
 
 export default async function takeScreenshot() {
+  const adbDir = await checkAdbExists();
   const now = moment(new Date()).format("DD-MM-YYYY-HH-mm-ss");
-  const adbDir = `${process.env.HOME}/Library/Android/sdk/platform-tools/adb`;
   const screenshotsDirectory = `${process.env.HOME}/adb-screenshots`;
   const filepath = `${screenshotsDirectory}/${now}.png`;
   if (!fs.existsSync(screenshotsDirectory)) {