+++ extensions/adb/src/adb-write.ts
@@ -1,12 +1,13 @@
 import { LaunchProps, showHUD } from "@raycast/api";
 import { execSync } from "child_process";
+import checkAdbExists from "./utils";
 
 interface AdbWriteArguments {
   text: string;
 }
 
 export default async function writeText(props: LaunchProps<{ arguments: AdbWriteArguments }>) {
-  const adbDir = `${process.env.HOME}/Library/Android/sdk/platform-tools/adb`;
+  const adbDir = await checkAdbExists();
   const text = props.arguments.text;
   await showHUD("✍️ Writing " + text);
   execSync(`${adbDir} shell input text '${text}'`);